Refactor: move JetpackAIQueryResponse out of rest client and adjust import statements
